Use darwinssl that works with keychain for darwin builds
Fixes issue with "Peer certificate cannot be authenticated with given CA certificates" when cross compiling for macos, since macos uses the keychain to store root certificates, and gnutls and libcurl do not recognize this, using darwinssl fixes the issue. Configuring the to use darwinssl when cross compiling can fail due to `/System/Library/Frameworks/Security.framework` not being recognized during cross compilation. Symlinking this to the root is a quick hack to get it to work, however the sysroot for the configure script should fix this.