Add `stream-file` command
`download-file` previously had an optional third argument for the path that
would default to stdout, however this would lead to confusion of the
incorrect number of arguments where entered. This brings back streaming
to stdout, as the third argument for path was made required for
`download-file`. STORJ_KEYPASS="password" environment variable can
be used before the command to make sure that it's able to decrypt the file.