Fix issue replacement pointer
If there is an error when retrieving a replacement pointer from the bridge, in this case an expired token, the status of the pointer wouldn't have been returned to it's previous state and the pointer would not have been replaced again. This solves the issue by returning the status of the pointer, and thus will be retried again for the replacement.