Reduce limit of pointers per request from 6 to 3
To help improve the timeout rate for this request. Setting a timeout to 20 seconds will often result in the client closing the connection before the request can be made, a timeout of 60 seconds does not have this issue. If we reduce the number of pointers that need to be requested, there are less requests that need to be made to the network.