Return download state
When implementing bindings such as in Ruby, allocating the download state within
storj_bridge_resolve_file will avoid the need to allocate from within Ruby in way
that Ruby garbage collection would be responsible for cleaning it up. In this case
the state is cleaned up within the function, so it should also be allocated with
the function.