Provide file metadata in storj_finished_upload_cb, not only file id
The bridge API returns complete file metadata for uploaded files, but
the storj_finished_upload_cb callback passed only the file id to
libstorj clients. Thus, clients needed to execute an additional call to
the bridge for retrieving metadata of interest like Created Time or
HMAC. This causes unnecessary complexity and latency for clients and
unnecessary load to the bridge.